去掉 REFER_FIELDS
