--- src/constants/fields.ts
@@ -1,10 +1,0 @@
-export const FIELD_MAPPER: TMapper = {
-  email: '邮箱',
-  name: '姓名',
-  experience: '工作经验',
-  intro: '个人简介',
-  phone: '电话',
-  leetCodeUrl: 'LeetCode链接',
-  thirdPersonIntro: '第三人称介绍',
-  resumeUrl: '个人简历'
-}
