+++ src/constants/referFields.ts
@@ -1,0 +1,15 @@
+export const REFER_FIELDS_MAPPER: TMapper = {
+  name: '姓名',
+  email: '邮箱',
+  phone: '电话',
+  experience: '工作经验',
+  resumeUrl: '简历链接',
+  intro: '个人简介',
+  referLinks: '内推链接',
+  thirdPersonIntro: '第三人称介绍',
+  leetCodeUrl: 'LeetCode链接',
+}
+
+export const REQUIRED_REFER_FIELD_VALUES = [
+  'name', 'email', 'experience', 'referLinks'
+]
