+++ src/views/RefereeRequest.vue
@@ -28,7 +28,7 @@
   import GetJobByIdGQL from '@/graphql/GetJobById.graphql'
   import GetResumeByIdGQL from '@/graphql/GetResumeById.graphql'
   import UpdateReferGQL from '@/graphql/UpdateRefer.graphql'
-  import {REFER_FIELDS_MAPPER} from "@/constants/refer"
+  import {REFER_FIELDS_MAPPER} from "@/constants/referFields"
   import {LEVEL_MAPPER} from "@/constants/level"
 
   @Component({
